requirements.txt
langchain
openai
pinecone-client
tiktoken
wikipedia
python-dotenv